refactor: relocate util function
